Refactor CSL-specific code out of spatial IR
